Fix model loading for blaze
