+++ face_alignment/detection/dlib/dlib_detector.py
@@ -1,45 +1,9 @@
-import sys
-import errno
 import warnings
-import os
 import cv2
 import dlib
 
-from urllib.parse import urlparse
-from torch.hub import download_url_to_file, get_dir, HASH_REGEX
-
 from ..core import FaceDetector
-
-# Pytorch load supports only pytorch models
-def load_file_from_url(url, model_dir=None, map_location=None, progress=True, check_hash=False, file_name=None):
-    if model_dir is None:
-        hub_dir = get_dir()
-        model_dir = os.path.join(hub_dir, 'checkpoints')
-
-    try:
-        os.makedirs(model_dir)
-    except OSError as e:
-        if e.errno == errno.EEXIST:
-            # Directory already exists, ignore.
-            pass
-        else:
-            # Unexpected OSError, re-raise.
-            raise
-
-    parts = urlparse(url)
-    filename = os.path.basename(parts.path)
-    if file_name is not None:
-        filename = file_name
-    cached_file = os.path.join(model_dir, filename)
-    if not os.path.exists(cached_file):
-        sys.stderr.write('Downloading: "{}" to {}\n'.format(url, cached_file))
-        hash_prefix = None
-        if check_hash:
-            r = HASH_REGEX.search(filename)  # r is Optional[Match[str]]
-            hash_prefix = r.group(1) if r else None
-        download_url_to_file(url, cached_file, hash_prefix, progress=progress)
-
-    return cached_file
+from ...utils import load_file_from_url
 
 class DlibDetector(FaceDetector):
     def __init__(self, device, path_to_detector=None, verbose=False):
